正しく計算されなかったので、unsigned long に変更。

+++ matmulHPC169.cl
@@ -1,6 +1,6 @@
 __attribute__((reqd_work_group_size(1,1,1)))
 __kernel void matMul(__global double *restrict a, __global const double *restrict b, __global const double *restrict c, const int N) {
-    unsigned int i, j, k, iter, iter2;
+    unsigned long i, j, k, iter, iter2;
     double sum;
 
     // for (i=0; i<N; i++){
@@ -17,7 +17,7 @@ __kernel void matMul(__global double *restrict a, __global const double *restric
     //     }
     // }
 
-    for (iter2=0; iter2<(unsigned int)N*N*N; iter2++){
+    for (iter2=0; iter2<(unsigned long)N*N*N; iter2++){
         i = iter2 / (N * N); iter = iter2 % (N * N);
         j = iter / N; k = iter % N;
         if(k == 0)
